Do not pass negative rows count to insert
In case of empty folder negative row number (-1) could be passed to insertRow model function within fetchMore function. Newer Qt versions seems to crash on this event.
This commit replaces -1 with 0 in this case fixing the issue. Inserting cannot be omited because model data should be invalidated to hide "+" sign on the empty folder

Fixes #1